+++ Accounts.Repository/Repository/ICivilEntitiesRpository.cs
@@ -11,9 +11,9 @@ namespace Accounts.Repository.Repository
     public interface ICivilEntitiesRpository
     {
         List<CivilEntity> GetAllCivilEntity();
-        CivilEntity FindCivilEntity(int id);
+        CivilEntity FindCivilEntity(long id);
         bool UpdateCivilEntity(VM_CivilEntity _VM_CivilEntity);
-        bool DeleteCivilEntity(int id);
+        bool DeleteCivilEntity(long id);
         bool AddACivilEntity(VM_CivilEntity _VM_CivilEntity);
     }
 }
